+++ packages/cli/src/commands/apply/constants.ts
@@ -1,2 +1,0 @@
-export const BEGIN_MARKER = 'WPK:BEGIN AUTO';
-export const END_MARKER = 'WPK:END AUTO';